requirements.txt
aldryn-bootstrap3==0.5.2
cmsplugin-filer==0.10.2
dj-database-url==0.3.0
django-appconf==1.0.1
django-classy-tags==0.6.2
django-cms==3.1.3
django-durationfield==0.5.2
django-filer==0.9.12
django-meta==0.3.1
django-meta-mixin==0.2.1
django-mptt==0.7.4
django-parler==1.5.1
django-polymorphic==0.7.2
django-reversion==1.8.5
django-sekizai==0.8.2
Django-Select2==4.3.2
django-taggit==0.17.1
django-taggit-autosuggest==0.2.7
django-taggit-templatetags==0.2.5
django-templatetag-sugar==1.0
django-treebeard==3.0
djangocms-admin-style==0.2.8
djangocms-blog==0.5.0
djangocms-column==1.5
djangocms-file==0.1
djangocms-flash==0.2.0
djangocms-googlemap==0.3
djangocms-inherit==0.1
djangocms-installer==0.8.1
djangocms-link==1.7.1
djangocms-picture==0.1
djangocms-style==1.5
djangocms-teaser==0.1
djangocms-text-ckeditor==2.6.0
djangocms-video==0.1
easy-thumbnails==2.2
html5lib==0.9999999
Pillow==3.0.0
pytz==2015.6
six==1.10.0
South==1.0.2
tzlocal==1.2
Unidecode==0.4.18
wheel==0.24.0
